backend/serializers.py
from rest_framework import serializers
from .models import ChatMessage

class ChatMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ChatMessage
        fields = ["id", "message", "response", "timestamp", "session_id"]
        read_only_fields = ["id", "response", "timestamp"]
