backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
openai>=1.40.0
python-dotenv>=1.0.1
